Generate a random request ID
The LifeOmic request processing uses the request ID attribute from
the APIG event payload as the default value for correlation IDs.
When an Alpha request is the root event in a tree of requests this
means that there is no correlation ID generated for the entire
tree.

Generating a request ID on Alpha requests will allow the LO
machinery to infer and propagate an correlation ID when Alpha is
the originator of the request.